[OSP-33530] Allow roles_data.yaml to have `./` prefix for checking fencing roles
When the tarball was created with:
`tar -cvzf tripleo-tarball-config.tar.gz ./*`

the files in the tarball have `./` prefix, which make
the check for the `roles_data.yaml` file to not match.
This change also checks if there is a `./roles_data.yaml`
as part of the tarball.

Jira: https://issues.redhat.com/browse/OSP-33530

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>