Make fields optional in v1beta1 to resolve reinstall issues
With 4.16 there is additional validation in olm when the installplan
gets deployed. In this validation the CR gets checked against the
old v1beta1 CRD, even the object storage version is v1beta2. As
there are no conversion webhooks in ospdo the verification fails.

To resolve this this adds the fields which get reported as errors
in v1beta1 as optional.

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>